Copied appointment related methods into AppointmentController
